[network][ip] Finish CIDR and subnet
- CIDR does not use class, making it more flexible for different
network sizes, VLSM variable length subnet mask, one address can match
mulitple CIDR prefixes (and router can be more efficient using that?)
- CIDR notation follow ip address with a slash and a decimal value `n`,
indicate the highest `n` bits are 1 for the routing mask, previously
known as network mask. i.e. 192.169.12.3/24 means the first 24 bits
should be 1 for the routing mask, the ip is 192.169.12.3, the network
is 192.169.12.0 the host identifier is 0.0.0.3, routing mask is
255.255.255.0